+++ csharp/SupermarketReceipt/Domain/Offers/Discount.cs
@@ -4,7 +4,7 @@ namespace SupermarketReceipt.Domain.Offers
 {
     public class Discount
     {
-        public Discount(Product product, string description, double discountAmount)
+        public Discount(Product product, string description, decimal discountAmount)
         {
             Product = product;
             Description = description;
@@ -12,7 +12,7 @@ namespace SupermarketReceipt.Domain.Offers
         }
 
         public string Description { get; }
-        public double DiscountAmount { get; }
+        public decimal DiscountAmount { get; }
         public Product Product { get; }
     }
 }
