+++ csharp/SupermarketReceipt/Domain/Prices/IPriceCatalog.cs
@@ -4,8 +4,8 @@ namespace SupermarketReceipt.Domain.Prices
 {
     public interface IPriceCatalog
     {
-        void AddProduct(Product product, double price);
+        void AddProduct(Product product, decimal price);
 
-        double GetUnitPrice(Product product);
+        decimal GetUnitPrice(Product product);
     }
 }
